requirements.txt
asgiref==3.8.1
certifi==2024.7.4
Django==4.1.13
djongo==1.3.6
dnspython==2.6.1
pymongo==4.8.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.2.4
